Learnings; Starting working by reading on-demand
While implementing a BTree approach, realised that there was no way
to keep a reasonably sized directory tree in memory, mostly because
each node must be sized to fit N files, and quickly run into having to
limit the amount of files per directory to a very low number.

This quickly grows the size of the stack allocation, which I expect to be
very low on embedded devices (e.g. esp32-c6). To prevent stack allocation
you can do an arena and work on indexes, but that runs into Rust language
itself, where keeping mutable statics is heavily discouraged.

Ultimately, the best approach is going to be to stop keeping things in
memory, and have operations read/write blocks on-demand.